retire doclayout
